K-means: switch pipeline to 16 bit.
Less RAM, less for loops in SIMD, simpler SIMD instructions.

On rtc_screen, speed 0, 50 runs, speed-ups are in %:

screen_recording_crd.1920_1080.y4m	0.483
screenshare_buganizer.1900_1306.y4m	0.718
screenshare_colorslides.1820_1320.y4m	0.913
screenshare_slidechanges.1850_1110.y4m	0.576
screenshare_youtube.1680_1178.y4m	0.465
slides_webplot.1920_1080.y4m	0.475
sc_web_browsing720p.y4m	0.690
screen_crd_colwinscroll.1920_1128.y4m	0.146
{OVERALL}	0.558

At speed 10:

screen_recording_crd.1920_1080.y4m	0.347
screenshare_buganizer.1900_1306.y4m	1.145
screenshare_colorslides.1820_1320.y4m	1.028
screenshare_slidechanges.1850_1110.y4m	0.945
screenshare_youtube.1680_1178.y4m	0.366
slides_webplot.1920_1080.y4m	0.999
sc_web_browsing720p.y4m	1.261
screen_crd_colwinscroll.1920_1128.y4m	0.133
{OVERALL}	0.778

+++ av1/encoder/block.h
@@ -326,7 +326,7 @@ typedef struct {
   //! The best color map found.
   uint8_t best_palette_color_map[MAX_PALETTE_SQUARE];
   //! A temporary buffer used for k-means clustering.
-  int kmeans_data_buf[2 * MAX_PALETTE_SQUARE];
+  int16_t kmeans_data_buf[2 * MAX_PALETTE_SQUARE];
 } PALETTE_BUFFER;
 
 /*! \brief Contains buffers used by av1_compound_type_rd()
